add_expense_group/requirements.txt
blinker==1.8.0
click==8.1.7
dnspython==2.6.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
passlib==1.7.4
gunicorn
pymongo==4.6.2
python-dotenv==1.0.1
python-http-client==3.3.7
sendgrid==6.11.0
starkbank-ecdsa==2.2.0
validators==0.23.2
Werkzeug==3.0.2
WTForms==3.1.2
requests==2.26.0
blinker==1.8.0
click==8.1.7
dnspython==2.6.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
passlib==1.7.4
pymongo==4.6.2
python-dotenv==1.0.1
python-http-client==3.3.7
sendgrid==6.11.0
starkbank-ecdsa==2.2.0
validators==0.23.2
Werkzeug==3.0.2
WTForms==3.1.2
